Apply MCUXpresso 2.11.0 release update
- Verified boards in scope:
    evkbimxrt1050
    evkbmimxrt1060
    evkmcimx7ulp
    evkmimx8mm
    evkmimx8mn
    evkmimx8mnddr3l
    evkmimx8mp
    evkmimx8mq
    evkmimxrt1010
    evkmimxrt1015
    evkmimxrt1020
    evkmimxrt1024
    evkmimxrt1060
    evkmimxrt1064
    evkmimxrt1160
    evkmimxrt1170
    evkmimxrt595
    evkmimxrt685
    frdmk32l2a4s
    frdmk32l2b
    frdmk32l3a6
    frdmk64f
    frdmk66f
    frdmke02z40m
    frdmke04z
    frdmke06z
    frdmke15z
    frdmke16z
    frdmke17z
    frdmkl27z
    frdmkv11z
    frdmkv31f
    hvpkv11z75m
    hvpkv31f120m
    lpc845breakout
    lpcxpresso51u68
    lpcxpresso54628
    lpcxpresso54s018
    lpcxpresso54s018m
    lpcxpresso55s06
    lpcxpresso55s06cp
    lpcxpresso55s16
    lpcxpresso55s28
    lpcxpresso55s69
    lpcxpresso802
    lpcxpresso804
    lpcxpresso824max
    lpcxpresso845max
    mimxrt685audevk
    twrkm34z50mv3
    twrkm34z75m
    twrkm35z75m

+++ evkmcimx7ulp/demo_apps/power_mode_switch/srtm/services/srtm_sensor_service.c
@@ -71,7 +71,7 @@ static srtm_status_t SRTM_SensorService_UpdateState(srtm_service_t service, srtm
 
     /* First allocate a notification and send to peer core */
     notif = SRTM_Notification_Create(handle->channel, SRTM_SENSOR_CATEGORY, SRTM_SENSOR_VERSION,
-                                     SRTM_SENSOR_NTF_STATE_CHANGED, sizeof(struct _srtm_sensor_payload));
+                                     SRTM_SENSOR_NTF_STATE_CHANGED, (uint16_t)sizeof(struct _srtm_sensor_payload));
     if (notif == NULL)
     {
         SRTM_DEBUG_MESSAGE(SRTM_DEBUG_VERBOSE_ERROR, "%s: alloc notification failed.\r\n", __func__);
@@ -98,7 +98,7 @@ static srtm_status_t SRTM_SensorService_ReportData(
 
     /* First allocate a notification and send to peer core */
     notif = SRTM_Notification_Create(handle->channel, SRTM_SENSOR_CATEGORY, SRTM_SENSOR_VERSION, SRTM_SENSOR_NTF_DATA,
-                                     sizeof(struct _srtm_sensor_payload));
+                                     (uint16_t)sizeof(struct _srtm_sensor_payload));
     if (!notif)
     {
         SRTM_DEBUG_MESSAGE(SRTM_DEBUG_VERBOSE_ERROR, "%s: alloc notification failed.\r\n", __func__);
@@ -140,7 +140,7 @@ static srtm_status_t SRTM_SensorService_Request(srtm_service_t service, srtm_req
     payloadLen = SRTM_CommMessage_GetPayloadLen(request);
 
     response = SRTM_Response_Create(channel, SRTM_SENSOR_CATEGORY, SRTM_SENSOR_VERSION, command,
-                                    sizeof(struct _srtm_sensor_payload));
+                                    (uint16_t)sizeof(struct _srtm_sensor_payload));
     if (!response)
     {
         return SRTM_Status_OutOfMemory;
